requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.3.0
pafy == 0.5.5
python_vlc == 3.0.16120
pytube == 12.1.0
